Remove all textures from PC Tower.
This is the only way for the PC to be able to load in the game.

We can add the texture back in manually for essential parts.

Back-up file is the .blend one in Discord chat. Don't blame me for this as I have test the PC without any textures and it's loaded. The fully textured one crashed completely.