Added 'Snapbox' when installing PC parts.
Removed the fully built PC from the 'Main' room as we got the separated parts done + the snapbox.

Added the GPU back in 'Main' room again because it got deleted somehow. Downscaled it already as well.

The snapbox might be not correctly scaled as I made it quickly enough to have something to do the function for.