Added global-use script for 'Main' room.
This is to preparing for the actual snapping system (low-effort implementing for sure) when player hold the components on hands and connect it with the snap area inside the PC case.

Correctly transform the hitbox for one of the PC parts - Fan, which somehow was not properly moved by accidentally.